20181012 create schedule controller

+++ app/controller/room.py
@@ -23,5 +23,6 @@ class RoomController:
 
     def delete_room_by_id(self, _id):
         room = self.get_room_by_id(_id)
-        db.session.delete(room)
-        db.session.commit()
+        if room:
+            db.session.delete(room)
+            db.session.commit()
